requirements.txt
azure-functions
requests

